pyproject.toml
[project]
name = "xhs-scraper"
version = "1.0.0"
description = "Xiaohongshu Content Analysis Pipeline"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "selenium>=4.15.0",
    "beautifulsoup4>=4.12.0",
    "opencv-python>=4.8.0",
    "ffmpeg-python>=0.2.0",
    "openai>=1.0.0",
    "jieba>=0.42.1",
    "textblob>=0.17.1",
    "jinja2>=3.1.2",
    "Pillow>=10.0.0",
    "rich>=13.7.0",
    "tqdm>=4.66.0",
    "scikit-learn>=1.3.0",
    "apify-client>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "ipython>=8.14.0",
    "jupyter>=1.0.0",
    "black>=23.0.0",
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]
viz = [
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[project.scripts]
xhs-pipeline = "xhs_pipeline:main"

[tool.uv]
dev-dependencies = [
    "ipython>=8.14.0",
    "jupyter>=1.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.black]
line-length = 120
target-version = ["py39"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"